rename: User -> Member 클래스명 변경
