+++ src/main/java/roomescape/domain/member/MemberName.java
@@ -1,15 +1,15 @@
-package roomescape.domain.user;
+package roomescape.domain.member;
 
 import java.util.regex.Pattern;
 
-public class UserName {
+public class MemberName {
     private static final Pattern NAME_FORMAT = Pattern.compile("^[a-zA-Z가-힣]*$");
     private static final int MIN_LENGTH = 2;
     private static final int MAX_LENGTH = 10;
 
     private final String name;
 
-    public UserName(String name) {
+    public MemberName(String name) {
         validateBlank(name);
         validatePattern(name);
         validateLength(name);
